+++ src/main/scala/io/scalac/auction/domain/LotActor.scala
@@ -1,10 +1,7 @@
-package io.scalac.auction
+package io.scalac.auction.domain
 
-import akka.actor.typed.ActorRef
-import akka.actor.typed.Behavior
-import akka.actor.typed.scaladsl.AbstractBehavior
-import akka.actor.typed.scaladsl.ActorContext
-import akka.actor.typed.scaladsl.Behaviors
+import akka.actor.typed.{ActorRef, Behavior}
+import akka.actor.typed.scaladsl.{AbstractBehavior, ActorContext, Behaviors}
 
 object LotActor {
