+++ src/main/scala/io/scalac/auction/domain/model/Auction.scala
@@ -1,3 +1,3 @@
-package io.scalac.auction.model
+package io.scalac.auction.domain.model
 
 case class Auction(id: String, status: AuctionStatus, lotIds: Seq[String])
