+++ src/main/scala/io/scalac/auction/domain/model/AuctionId.scala
@@ -1,3 +1,3 @@
-package io.scalac.auction.model
+package io.scalac.auction.domain.model
 
 case class AuctionId(id: String) extends AnyVal
