+++ src/main/scala/io/scalac/auction/domain/model/Lot.scala
@@ -1,3 +1,3 @@
-package io.scalac.auction.model
+package io.scalac.auction.domain.model
 
 case class Lot(id: String, auctionId: String, description: Option[String], topBidder: Option[String], topBid: Option[BigDecimal])
