+++ src/main/scala/io/scalac/auction/domain/model/LotId.scala
@@ -1,3 +1,3 @@
-package io.scalac.auction.model
+package io.scalac.auction.domain.model
 
 case class LotId(id: String) extends AnyVal
