+++ src/test/scala/io/scalac/auction/AuctionServiceSpec.scala
@@ -1,8 +1,9 @@
 package io.scalac.auction
 
 import akka.actor.testkit.typed.scaladsl.ActorTestKit
-import io.scalac.auction.model.ServiceFailure.{AuctionNotReady}
-import io.scalac.auction.model.{Auction, AuctionId, AuctionStates, Lot, ServiceFailure}
+import io.scalac.auction.domain.{AuctionActorManager, DefaultAuctionService}
+import io.scalac.auction.domain.model.ServiceFailure.AuctionNotReady
+import io.scalac.auction.domain.model.{Auction, AuctionId, AuctionStates, Lot, ServiceFailure}
 import io.scalac.util.{Configs, ExecutionContexts}
 import org.scalamock.scalatest.AsyncMockFactory
 import org.scalatest.BeforeAndAfterAll
